[backend] Cleaning and Documenting
Added more documentation and separated some functions/endpoints
