+++ backend/decorators/__init__.py
@@ -1,3 +1,10 @@
-from .csrf import csrf
+"""
+Package for decorators.
 
-__all__ = ["csrf"]
+Decorators:
+  - csrf_protected
+"""
+
+from .csrf_protection import csrf_protected
+
+__all__ = ["csrf_protected"]
