--- backend/decorators/authorization.py
@@ -1,6 +1,0 @@
-from functools import wraps
-
-
-def requires_authorization():
-    # TODO: Implement decorator
-    pass
