+++ backend/routes/student_routes.py
@@ -1,34 +1,33 @@
-from datetime import timedelta
-from http import HTTPStatus
+"""
+Student Routes (Protected)
+API Endpoint: '/api/v1/students'
+"""
+
 import json
+from http import HTTPStatus
 from typing import Any
 from flask import Blueprint, jsonify, make_response, request
 from flask_jwt_extended import (
     create_access_token,
-    create_refresh_token,
     get_jwt,
     get_jwt_identity,
     jwt_required,
     current_user,
     set_access_cookies,
-    set_refresh_cookies,
     unset_jwt_cookies,
 )
-from decorators import csrf
+from decorators import csrf_protected
 from models.committees.committee import Committee
 from models.committees.committee_position import CommitteePosition
 from models.core.student import Student, StudentMembership
-from services.core.student import login, assign_password, get_permissions
-from utility.gc import delete_file, upload_file
+from services.core.student import login, assign_password, get_permissions, update
 from utility.translation import retrieve_languages
-from werkzeug.security import check_password_hash
-from utility.database import db
 
 student_bp = Blueprint("student", __name__)
 
 
 @student_bp.route("/login", methods=["POST"])
-@csrf
+@csrf_protected
 def student_login():
     data = request.get_json()
 
@@ -41,7 +40,7 @@ def student_login():
 
 
 @student_bp.route("/", methods=["PUT"])
-@csrf
+@csrf_protected
 @jwt_required(refresh=True)
 def update_student():
     student_id = get_jwt_identity()
@@ -50,59 +49,10 @@ def update_student():
     if not student or not isinstance(student, Student):
         return jsonify({"error": "Invalid credentials"}), 401
 
-    profile_picture = request.files.get("profile_picture")
-    currentPassword = request.form.get("current_password")
-    newPassword = request.form.get("new_password")
-
-    if not currentPassword:
-        return jsonify({"error": "No current password provided"}), 400
-
-    if not check_password_hash(getattr(student, "password_hash"), currentPassword):
-        return jsonify({"error": "Invalid current password"}), 400
-
-    file_extension = profile_picture.filename.split(".")[-1]
-
-    if profile_picture:
-        # Delete previous image if it exists
-        if getattr(student, "profile_picture_url"):
-            delete_file(
-                getattr(student, "profile_picture_url"),
-            )
-
-        result = upload_file(
-            file=profile_picture,
-            file_name=f"{student_id}.{file_extension}",
-            path="profile",
-        )
-
-        setattr(student, "profile_picture_url", result)
-
-    if newPassword:
-        result = assign_password({"email": student.email, "password": newPassword})
-        if not result:
-            return jsonify({"error": "Failed to update password"}), 500
-
-    db.session.commit()
-    response = make_response()
-    permissions_and_role = get_permissions(getattr(student, "student_id"))
-    additional_claims = {
-        "role": permissions_and_role.get("role"),
-        "permissions": permissions_and_role.get("permissions"),
-    }
-    access_token = create_access_token(
-        identity=student, fresh=True, additional_claims=additional_claims
+    return update(
+        request=request,
+        student=student,
     )
-    refresh_token = create_refresh_token(
-        identity=student, expires_delta=timedelta(days=30)
-    )
-    set_access_cookies(
-        response=response, encoded_access_token=access_token, max_age=timedelta(hours=1)
-    )
-
-    set_refresh_cookies(response=response, encoded_refresh_token=refresh_token)
-
-    response.status_code = HTTPStatus.OK
-    return response
 
 
 @student_bp.route("/refresh", methods=["POST"])
@@ -136,7 +86,7 @@ def get_student_permissions():
     return get_permissions(student_id)
 
 
-# TODO: Remove Later?
+# TODO: Remove Later
 @student_bp.route(rule="/reset", methods=["POST"])
 def reset_password():
     data = request.get_json()
@@ -155,7 +105,7 @@ def reset_password():
 
 
 @student_bp.route("/me", methods=["GET"])
-@jwt_required(fresh=True)
+@jwt_required()
 def get_student_callback():
     language_code = retrieve_languages(request.args)
     student_id = get_jwt_identity()
