+++ backend/services/core/__init__.py
@@ -1,4 +1,16 @@
-from .student import login, change_password, assign_password, get_permissions
+from .student import (
+    login,
+    change_password,
+    assign_password,
+    get_permissions,
+    update,
+)
 
 
-__all__ = ["login", "change_password", "assign_password", "get_permissions"]
+__all__ = [
+    "login",
+    "change_password",
+    "assign_password",
+    "get_permissions",
+    "update",
+]
