+++ backend/services/core/student.py
@@ -1,5 +1,6 @@
 from datetime import timedelta
-from flask import jsonify, make_response
+from http import HTTPStatus
+from flask import Request, Response, jsonify, make_response
 from werkzeug.security import check_password_hash, generate_password_hash
 from flask_jwt_extended import (
     create_access_token,
@@ -12,16 +13,18 @@ from models.content.author import Author, AuthorType
 from models.core.permissions import StudentPermission
 from models.core.student import Student
 from utility.database import db
+from utility.gc import delete_file, upload_file
 
 
 def login(data: Dict[str, Any]):
     """
     Login function that validates the user's credentials and generates an access token.
 
-        Parameters:
-            data (Dict[str, Any]): A dictionary containing user data with "email" and "password" keys.
-        Returns:
-            dict: A dictionary containing the "access_token" if the login is successful, otherwise None.
+    Args:
+        data (Dict[str, Any]): A dictionary containing user data with "email" and "password" keys.
+
+    Returns:
+        dict: A dictionary containing the "access_token" if the login is successful, otherwise None.
 
     """
     email = data.get("email")
@@ -121,6 +124,75 @@ def assign_password(data: Dict[str, Any]):
     return True
 
 
+def update(request: Request, student: Student) -> Response:
+    """
+    Updates the student's profile picture and password.
+
+    Args:
+        request (Request): The request object.
+        student (Student): The student object.
+
+    Returns:
+        response (Response): The response object.
+
+    """
+
+    profile_picture = request.files.get("profile_picture")
+    currentPassword = request.form.get("current_password")
+    newPassword = request.form.get("new_password")
+
+    if not currentPassword:
+        return jsonify({"error": "No current password provided"}), 400
+
+    if not check_password_hash(getattr(student, "password_hash"), currentPassword):
+        return jsonify({"error": "Invalid current password"}), 400
+
+    file_extension = profile_picture.filename.split(".")[-1]
+
+    if profile_picture:
+        # Delete previous image if it exists
+        if getattr(student, "profile_picture_url"):
+            delete_file(
+                getattr(student, "profile_picture_url"),
+            )
+
+        result = upload_file(
+            file=profile_picture,
+            file_name=f"{student.student_id}.{file_extension}",
+            path="profile",
+        )
+
+        setattr(student, "profile_picture_url", result)
+
+    if newPassword:
+        result = assign_password({"email": student.email, "password": newPassword})
+        if not result:
+            return jsonify({"error": "Failed to update password"}), 500
+
+    db.session.commit()
+    response = make_response()
+    permissions_and_role = get_permissions(getattr(student, "student_id"))
+    additional_claims = {
+        "role": permissions_and_role.get("role"),
+        "permissions": permissions_and_role.get("permissions"),
+    }
+    access_token = create_access_token(
+        identity=student, fresh=True, additional_claims=additional_claims
+    )
+    refresh_token = create_refresh_token(
+        identity=student, expires_delta=timedelta(days=30)
+    )
+    set_access_cookies(
+        response=response, encoded_access_token=access_token, max_age=timedelta(hours=1)
+    )
+
+    set_refresh_cookies(response=response, encoded_refresh_token=refresh_token)
+
+    response.status_code = HTTPStatus.OK
+
+    return response
+
+
 def get_permissions(student_id: str) -> Dict[str, Any]:
     all_permissions_and_role = {
         "role": None,
@@ -157,7 +229,7 @@ def get_permissions(student_id: str) -> Dict[str, Any]:
 
 
 def get_student(token: str):
-    if token is None:
+    if token == "" or token is None:
         return None
 
     return Student.query.filter_by(student_id=token).first()
